Create CONTRIBUTING.md
Added CONTRIBUTING.md file .The aim is to create a welcoming and accessible entry point for new contributors by guiding them through the process of contributing to the project. This section provides a step-by-step overview of how to fork, clone, and create branches, along with instructions for making, committing, and pushing changes, and ultimately, submitting a pull request.